creating a utils module for h5 utils stuff
